refactor(api): Move db services to separate module
